+++ Source/Core/oneTimeWarning.js
@@ -7,7 +7,7 @@ define([
         defaultValue,
         defined,
         DeveloperError) {
-    "use strict";
+    'use strict';
 
     var warnings = {};