+++ Tools/eslint-config-cesium/CHANGES.md
@@ -10,6 +10,7 @@ Change Log
 * Enable [no-loop-func](http://eslint.org/docs/rules/no-loop-func).
 * Enable [no-undef-init](http://eslint.org/docs/rules/no-undef-init).
 * Enable [no-implicit-globals](http://eslint.org/docs/rules/no-implicit-globals).
+* Enable [quotes](http://eslint.org/docs/rules/quotes) to enforce use of single quotes.
 * Enable [no-trailing-spaces](http://eslint.org/docs/rules/no-trailing-spaces).
 * Enable [no-lonely-if](http://eslint.org/docs/rules/no-lonely-if).
 * Enable [no-unused-expressions](http://eslint.org/docs/rules/no-unused-expressions).